+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/extradata/reportingdata/State.java
@@ -18,6 +18,35 @@ public class State{
     this.stateType = stateType;
   }
 
-  
+  /**
+   * @return the priority order if it's considered as a validation state, null otherwise
+   */
+  private Integer getPriorityOrder(){
+    return switch (stateType){
+      case "PARTIELINT" -> 3;
+      case "VALPAP" -> 2;
+      case "VALINT" -> 1;
+      default -> null;
+    };
+  }
+
+  /**
+   * @return true if it's considered as a validation state
+   */
+  public boolean isValidationState(){
+    return (getPriorityOrder() != null);
+  }
+
+  public boolean isPriorTo(State otherState){
+    if(otherState == null) {
+      return true;
+    }
+    return (
+            // True if this state is more priority than the other
+            this.getPriorityOrder() < otherState.getPriorityOrder()
+            //True if the other state has the same priority and this state is more recent
+            || (this.getPriorityOrder().equals(otherState.getPriorityOrder()) && this.getTimestamp() > otherState.getTimestamp())
+    );
+  }
   
 }